+++ docs/conf.py
@@ -123,7 +123,7 @@ html_theme = 'sphinx_rtd_theme'
 # further.  For a list of options available for each theme, see the
 # documentation.
 html_theme_options = {
-    'analytics_id': 'UA-52965619-6',
+    'analytics_id': 'UA-52965619-6'
 }
 
 # Add any paths that contain custom themes here, relative to this directory.
@@ -305,3 +305,4 @@ def setup(app):
             substitutions=substitutions
         )
     )
+    utils.patch_theme_template(app, templates_path=templates_path[0])