+++ HISTORY.rst
@@ -173,6 +173,7 @@
   argument: "redis_url". Example: redis://127.0.0.1:6379/0
 - Add the setting LOCK_MANAGER_BACKEND_ARGUMENTS.
 - Automate documentation building dependencies.
+- Add sphinx sitemap extension.
 
 3.3.11 (2019-XX-XX)
 ===================