+++ HISTORY.rst
@@ -1,6 +1,8 @@
 3.2.11 (2019-XX-XX)
 ===================
 - Backport transaction handling to document model events.
+- Update example LDAP authentication settings file.
+- Update FAQ entry about the LDAP file.
 
 3.2.10 (2019-11-19)
 ===================