+++ HISTORY.rst
@@ -176,6 +176,7 @@
 - Add sphinx sitemap extension.
 - Move the file patching code from the Dependency class to a
   generalized utility of the storages app.
+- Add book link to the documentation.
 
 3.3.11 (2019-XX-XX)
 ===================