+++ HISTORY.rst
@@ -7,6 +7,7 @@
 - Add sphinx sitemap extension.
 - Move the file patching code from the Dependency class to a
   generalized utility of the storages app.
+- Add book link to the documentation.
 
 3.2.10 (2019-11-19)
 ===================